moves randoDino into its own folder
